removed lots of useless code
