+++ .gitignore
@@ -31,6 +31,11 @@ airbyte-integrations/connectors/**/src/test/resources/debug_resources
 # Connector CAT test related files
 airbyte-integrations/connectors/**/integration_tests/temp
 
+# Connector local built artefacts
+airbyte-integrations/connectors/**/bin
+airbyte-integrations/connectors/**/lib
+airbyte-integrations/connectors/**/lib64
+
 # Python
 *.egg-info
 __pycache__
@@ -112,3 +117,5 @@ target/
 # CDK load artifacts
 airbyte-cdk/bulk/core/load/airbyte-cdk-load/
 airbyte-integrations/connectors/**/airbyte-cdk-load/
+
+
