+++ airbyte-integrations/connectors/destination-mariadb/destination_mariadb/destination.py
@@ -1,14 +1,6 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-import tempfile
 from logging import Logger
-from pathlib import Path
 from typing import Any, Iterable, Mapping, Optional
 
-import pydevd_pycharm
 from airbyte.secrets import SecretString
 from airbyte.strategies import WriteStrategy
 from airbyte_cdk.destinations import Destination
@@ -19,7 +11,6 @@ from airbyte_cdk.models import (
     ConnectorSpecification,
     DestinationSyncMode,
     Status,
-    Type,
 )
 
 import logging
@@ -28,8 +19,6 @@ logger = logging.getLogger("airbyte")
 from destination_mariadb import mariadb_processor
 from destination_mariadb.common.catalog.catalog_providers import CatalogProvider
 from destination_mariadb.config import ConfigModel
-# import pydevd_pycharm
-BATCH_SIZE = 150
 
 
 class DestinationMariaDB(Destination):
@@ -43,15 +32,12 @@ class DestinationMariaDB(Destination):
                 host=config.indexing.host,
                 port=config.indexing.port,
                 database=config.indexing.database,
-                # schema_name=config.indexing.default_schema,
                 username=config.indexing.username,
                 password=SecretString(config.indexing.credentials.password),
             ),
             splitter_config=config.processing,
-            embedder_config=config.embedding,  # type: ignore [arg-type]  # No common base class
+            embedder_config=config.embedding,
             catalog_provider=CatalogProvider(configured_catalog),
-            #temp_dir=Path(tempfile.mkdtemp()),
-            #temp_file_cleanup=True,
         )
 
 
@@ -62,7 +48,8 @@ class DestinationMariaDB(Destination):
         configured_catalog: ConfiguredAirbyteCatalog,
         input_messages: Iterable[AirbyteMessage],
     ) -> Iterable[AirbyteMessage]:
-        #pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
+        # import pydevd_pycharm
+        # pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
         logger.info("Write starting up")
         parsed_config = ConfigModel.parse_obj(config)
         self._init_sql_processor(config=parsed_config, configured_catalog=configured_catalog)
@@ -78,18 +65,17 @@ class DestinationMariaDB(Destination):
         logger.info("After this, I should terminate")
 
 
-    def check(self, logger: Logger, config: Mapping[str, Any]) -> AirbyteConnectionStatus:
-        _ = logger  # Unused
-
+    def check(self, check_logger: Logger, config: Mapping[str, Any]) -> AirbyteConnectionStatus:
         #pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
 
         try:
             parsed_config = ConfigModel.parse_obj(config)
             self._init_sql_processor(config=parsed_config)
-            foo = self.sql_processor.sql_config.get_sql_engine().connect()
-            foo.close()
+            self.sql_processor.sql_config.get_sql_engine().connect()
+            check_logger.info("Connection successful")
             return AirbyteConnectionStatus(status=Status.SUCCEEDED)
         except Exception as e:
+            check_logger.error(f"An exception occurred: {repr(e)}")
             return AirbyteConnectionStatus(
                 status=Status.FAILED, message=f"An exception occurred: {repr(e)}"
             )
