+++ airbyte-integrations/connectors/destination-mariadb/pyvenv.cfg
@@ -1,0 +1,5 @@
+home = /usr/bin
+include-system-site-packages = false
+version = 3.11.11
+executable = /usr/bin/python3.11
+command = /usr/bin/python3.11 -m venv /home/alexander/Code/airbyte/airbyte/airbyte-integrations/connectors/destination-mariadb
